requirements.txt
fastapi
uvicorn
pytest
requests
wonderwords
aiofiles
pillow